Silence make "not found" errors
When recursing below engine_atecc/cryptoauthlib/lib, the stdout output
of the make command is executed which results in errors such as:

    /bin/sh: 1: make[3]:: not found
    /bin/sh: 1: make[4]:: not found

These errors are ignored because the recipe is prefixed with a dash.
As a result, any compilation errors that may occur in one of the
subdirectories will be ignored as well.

Copy the recursion recipe used in the other Makefiles, thereby silencing
the make errors and allowing legitimate compiler errors to stop the
build.

Signed-off-by: Lukas Wunner <lukas@wunner.de>